Split PgStore into its own file, add doc comments
