+++ src/lib.rs
@@ -1,3 +1,7 @@
+//! Event store for working with event-store-driven applications
+
+#![deny(missing_docs)]
+
 extern crate fallible_iterator;
 extern crate postgres;
 #[macro_use]
@@ -5,93 +9,94 @@ extern crate serde_derive;
 extern crate serde;
 extern crate serde_json;
 
+pub mod pg;
 pub mod testhelpers;
 
-use fallible_iterator::FallibleIterator;
-use postgres::types::ToSql;
-use postgres::Connection;
-use serde::de::DeserializeOwned;
-use serde::Deserialize;
-use serde_json::{from_value, Value as JsonValue};
 use std::fmt::Debug;
-use std::marker::PhantomData;
 
+/// Trait to be implemented by all domain events
 pub trait Event {}
+
+/// Trait to be implemented by the enum of all domain events
+///
+/// ```rust
+/// # use event_store_rs::Events;
+/// struct EventA;
+/// struct EventB;
+///
+/// enum DomainEvents {
+///     A(EventA),
+///     B(EventB),
+/// }
+///
+/// impl Events for DomainEvents {}
+/// ```
 pub trait Events {}
+
+/// A query to be passed to the store
+///
+/// This trait must be implemented for whichever type you want to pass to a particular store. See
+/// impls below for examples.
 pub trait StoreQuery {}
+
+/// Aggregator trait
+///
+/// This takes three type items:
+///
+/// * `E: Events` – The enum of domain events that rows from the backing store will be parsed to
+/// * `A` – The type of the query args to use when querying the backing store. Can be as simple as
+/// a `String`, but using a `struct` is recommended for readability
+/// * `Q: StoreQuery` – The query object to pass to the backing store. It should be built from `A`
+///
+/// `Aggregator` has trait bounds of `Copy + Clone + Debug + Default`. All can be `derive()`d easily
+/// except `Default`, but that's easy enough to implement. `Default` should be the initial state
+/// of the entity before the events are reduced onto it. Example:
+///
+/// ```rust
+/// #[derive(Clone, Debug)]
+/// struct ExampleUser {
+///     name: String,
+///     email: String,
+///     bio: Option<String>
+/// }
+///
+/// impl Default for ExampleUser {
+///     fn default() -> Self {
+///         Self {
+///             name: "".into(),
+///             email: "".into(),
+///             bio: None,
+///         }
+///     }
+/// }
+/// ```
 pub trait Aggregator<E: Events, A, Q: StoreQuery>: Copy + Clone + Debug + Default {
+    /// Apply an event `E` to `acc`, returning a copy of `Self` with updated fields. Can also just
+    /// return `acc` if nothing has changed.
     fn apply_event(acc: Self, event: &E) -> Self;
 
+    /// Produce a query object from some query arguments
     fn query(field: A) -> Q;
 }
 
-pub struct PgQuery<'a> {
-    query: &'a str,
-    args: Vec<Box<ToSql>>,
-}
-
-impl<'a> StoreQuery for PgQuery<'a> {}
-
-impl<'a> PgQuery<'a> {
-    pub fn new(query: &'a str, args: Vec<Box<ToSql>>) -> Self {
-        Self { query, args }
-    }
-}
-
+/// Store trait
+///
+/// Backing stores must implement this trait to maintain portability. Additional bounds can be
+/// added to `E: Events`. For example, the Postgres store implements `Store` with
+/// `Events: Events + DeserializeOwned` so that the event data can be deserialized by Serde:
+///
+/// ```ignore
+/// impl<'a, E> Store<E, PgQuery<'a>> for PgStore<E>
+/// where
+///     E: Events + DeserializeOwned,
+/// {
+///     // ...
+/// }
+/// ```
 pub trait Store<E: Events, Q: StoreQuery> {
+    /// Query the backing store and return an entity `T`, reduced from queried events
     fn aggregate<T, A>(&self, query: A) -> T
     where
         E: Events,
         T: Aggregator<E, A, Q>;
-}
-
-pub struct PgStore<E: Events> {
-    phantom: PhantomData<E>,
-    conn: Connection,
-}
-
-impl<'a, E> PgStore<E>
-where
-    E: Events + Deserialize<'a>,
-{
-    pub fn new(conn: Connection) -> Self {
-        Self {
-            phantom: PhantomData,
-            conn,
-        }
-    }
-}
-
-impl<'a, E> Store<E, PgQuery<'a>> for PgStore<E>
-where
-    E: Events + DeserializeOwned,
-{
-    fn aggregate<T, A>(&self, query_args: A) -> T
-    where
-        T: Aggregator<E, A, PgQuery<'a>>,
-    {
-        let PgQuery { query, args } = T::query(query_args);
-
-        let mut params: Vec<&ToSql> = Vec::new();
-
-        for (i, _arg) in args.iter().enumerate() {
-            params.push(&*args[i]);
-        }
-
-        let trans = self.conn.transaction().expect("Tranny");
-        let stmt = trans.prepare(&query).expect("Prep");
-
-        let results = stmt
-            .lazy_query(&trans, &params, 1000)
-            .expect("Query")
-            .map(|row| {
-                let json: JsonValue = row.get("data");
-                let evt: E = from_value(json).expect("Decode");
-
-                evt
-            }).fold(T::default(), |acc, event| T::apply_event(acc, &event))
-            .expect("Fold");
-
-        results
-    }
 }
